pyproject.toml
[project]
name = "mep-predict-2d"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "natsort>=8.4.0",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]
